+++ src/main/java/duke/Storage.java
@@ -1,12 +1,17 @@
 package duke;
 
+import duke.task.Deadline;
+import duke.task.Event;
 import duke.task.Task;
+import duke.task.ToDo;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Scanner;
 
@@ -55,7 +60,7 @@ public class Storage {
 
             while (scanner.hasNextLine()) {
                 String dataLine = scanner.nextLine();
-                Task task = Parser.dataToTask(dataLine);
+                Task task = dataToTask(dataLine);
                 taskList.add(task);
             }
 
@@ -78,7 +83,7 @@ public class Storage {
 
             for (Task task : taskList) {
                 // Write each task with specified format
-                writer.write(Parser.taskData(task) + "\n");
+                writer.write(taskData(task) + "\n");
             }
 
             writer.close();
@@ -87,5 +92,63 @@ public class Storage {
         }
     }
 
+    /**
+     * Converts a Task object into a data string for storage.
+     *
+     * @param task Task to convert.
+     * @return Data string representing the Task.
+     */
+    public static String taskData(Task task) {
+        String taskType = task instanceof ToDo ? "T" : task instanceof Deadline ? "D" : "E";
+        String status = task.getIsDone() ? "1" : "0";
+        String description = task.getDescription();
+
+        StringBuilder data = new StringBuilder();
+        data.append(taskType).append(" | ").append(status).append(" | ").append(description);
+
+        // Add additional information for Deadline and Event tasks
+        if (task instanceof Deadline) {
+            LocalDateTime deadlineDateTime = ((Deadline) task).getBy();
+            String formattedDateTime = deadlineDateTime.format(DateTimeFormatter
+                    .ofPattern("dd/MM/yyyy HHmm"));
+            data.append(" | ").append(formattedDateTime);
+        } else if (task instanceof Event) {
+            data.append(" | ").append(((Event) task).getFrom()).append(" | ")
+                    .append(((Event) task).getTo());
+        }
+
+        return data.toString();
+    }
+
+    /**
+     * Parses a data string and creates a Task from it.
+     *
+     * @param data Data string to parse.
+     * @return A Task object parsed from the data string.
+     */
+    public static Task dataToTask(String data) {
+        String[] info = data.split(" \\| ");
+        String taskType = info[0];
+        String status = info[1];
+        String description = info[2];
+
+        Task task;
+
+        if (taskType.equals("T")) {
+            task = new ToDo(description);
+        } else if (taskType.equals("D")) {
+            task = new Deadline(description, info[3]);
+        } else if (taskType.equals("E")) {
+            task = new Event(description, info[3], info[4]);
+        } else {
+            throw new IllegalArgumentException("☹ OOPS!!! Invalid task in data");
+        }
+
+        if (status.equals("1")) {
+            task.markAsDone();
+        }
+
+        return task;
+    }
 
 }
