+++ src/test/java/duke/ParserTest.java
@@ -1,61 +1,95 @@
 package duke;
 
+import duke.exception.InvalidDescriptionException;
 import duke.task.Deadline;
 import duke.task.Event;
-import duke.task.Task;
-import java.time.LocalDateTime;
+import duke.task.TaskList;
+import duke.task.ToDo;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.Test;
 
 public class ParserTest {
-    @Test
-    public void testTaskToStringData() {
-        Task task = new Deadline("math homework", "12/12/2002 1700");
-        String taskString = Parser.taskData(task);
-        String expected = "D | 0 | math homework | 12/12/2002 1700";
+    private Parser parser = new Parser();
+    private TaskList taskList = new TaskList();
 
-        assertEquals(expected, taskString);
+    @Test
+    public void testAddToDo() {
+        try {
+            parser.addToDo("todo watch videos", taskList);
+            assertEquals(1, taskList.getSize());
+        } catch (Exception e) {
+            fail("Exception should not be thrown.");
+        }
     }
 
     @Test
-    public void testStringDataToEventTask() {
-        String taskString = "E | 1 | project meeting | Mon 2pm | 6pm";
-        Task task = Parser.dataToTask(taskString);
-
-        assertEquals(Event.class, task.getClass());
-
-        assertEquals("project meeting", task.getDescription());
-
-        assertTrue(task.getIsDone());
-
-        Event event = (Event) task;
-        assertEquals("Mon 2pm", event.getFrom());
-        assertEquals("6pm", event.getTo());
+    public void testAddDeadline() {
+        try {
+            parser.addDeadline("deadline assignment /by 12/03/2040 1345", taskList);
+            assertEquals(1, taskList.getSize());
+        } catch (Exception e) {
+            fail("Exception should not be thrown.");
+        }
     }
 
     @Test
-    public void testStringToDeadlineTask() {
-        String taskString = "D | 0 | assignment 0 | 12/12/2002 1700";
-        Task task = Parser.dataToTask(taskString);
-
-        assertEquals(Deadline.class, task.getClass());
-
-        assertEquals("assignment 0", task.getDescription());
-
-        assertTrue(!task.getIsDone());
-
-        Deadline deadline = (Deadline) task;
-        assertEquals(LocalDateTime.parse("2002-12-12T17:00"), deadline.getBy());
+    public void testAddEvent() {
+        try {
+            parser.addEvent("event meeting /from 6th June /to 9th June", taskList);
+            assertEquals(1, taskList.getSize());
+        } catch (Exception e) {
+            fail("Exception should not be thrown.");
+        }
     }
 
-    // Supposed to be Deadline test
     @Test
-    public void testDeadlineString() {
-        Task task = new Deadline("math homework", "12/12/2002 1700");
-        String expected = "[D][ ] math homework (by: 12 Dec 2002 5:00 PM)";
+    public void testRemoveTask() {
+        try {
+            parser.addToDo("todo play!", taskList);
 
-        assertEquals(expected, task.toString());
+            parser.deleteTask("delete 1", taskList);
+            assertEquals(0, taskList.getSize());
+        } catch (Exception e) {
+            fail("Exception should not be thrown.");
+        }
+    }
+
+    @Test
+    public void testFindMatchingTaskList() throws Exception {
+        taskList.addTask(new ToDo("return book"));
+        taskList.addTask(new Deadline("book review", "07/12/2019 1920"));
+        taskList.addTask(new Event("meeting", "Mon 6pm", "10pm"));
+
+        TaskList matchingTaskList = parser.findMatchingTaskList("find book", taskList);
+
+        assertEquals(2, matchingTaskList.getSize());
+        assertEquals("[T][ ] return book", matchingTaskList.getTask(0).toString());
+        assertEquals("[D][ ] book review (by: 07 Dec 2019 7:20 PM)",
+                matchingTaskList.getTask(1).toString());
+    }
+
+    @Test
+    public void testFindNoMatchingTaskList() throws Exception {
+        taskList.addTask(new ToDo("return book"));
+        taskList.addTask(new Deadline("book review", "07/12/2019 1920"));
+        taskList.addTask(new Event("meeting", "Mon 6pm", "10pm"));
+
+        TaskList matchingTaskList = parser.findMatchingTaskList("find hello", taskList);
+
+        assertEquals(0, matchingTaskList.getSize());
+    }
+
+    @Test
+    public void testFindMatchingTaskListEmptyKeyword() throws Exception {
+        taskList.addTask(new ToDo("return book"));
+        taskList.addTask(new Deadline("book review", "07/12/2019 1920"));
+        taskList.addTask(new Event("meeting", "Mon 6pm", "10pm"));
+
+        assertThrows(InvalidDescriptionException.class, () -> {
+            parser.findMatchingTaskList("find ", taskList);
+        });
     }
 
 }
